+++ gammapy/estimators/map/__init__.py
@@ -1,3 +1,4 @@
 from .asmooth import *
 from .excess import *
 from .ts import *
+from .core import *
