+++ gammapy/estimators/map/core.py
