+++ gammapy/estimators/map/tests/test_core.py
